Add option to filter guild achievements
This bug is rather interesting. While the API no longer returns the
actual faction of the guild, there still seems to be the primary faction
for every guild, if anything, it's entirely possible that under the hood
there's 2 guilds, one for horde members and one for ally, but they're
presented as one to users. The guild achievement spam is caused by
achievements of the secondary faction, so in a horde guild people will
be spammed with ally achievements and vice versa. For whatever reason
they get reset regularly, and whenever someone logs in they pop.
Ref: #59, #68